+++ auth-server/src/main/java/com/nycloud/auth/service/SysUserService.java
@@ -13,5 +13,5 @@ public interface SysUserService {
 
     SysUser findByUserName(String username);
 
-    SysUser selectUserResources(Integer userId);
+    SysUser selectUserResources(Long userId);
 }
